requirements.txt
Pillow==9.0.1
PyPDF2==1.26.0

